// src/app/modules/models/DetailedEmployee.ts
    export interface DetailedEmployee {
        tenentId: number;
        locationId: number;
        employeeId: string;
        contractType: string;
        pfid: string;
        subscribedDate: Date;
        agreedSubAmount: number;
        reSubscribed: Date;
        employeeType: string;
        arabicName: string;
        englishName: string;
        jobTitleCode: number;
        jobTitleName: string;
        department: number;
        departmentName: string;
        empGender: number;
        empBirthday: Date;
        empMaritalStatus: number;
        salary: number;
        empWorkTelephone: string;
        empWorkEmail: string;
        mobileNumber: string;
        next2KinName: string;
        next2KinMobNumber: string;
        nationCode: number;
        nationName: string;
        empCidNum: string;
        empPaciNum: string;
        empOtherId: string;
        empStatus: number;
        dateOfJoining: Date;
        endDate: Date;
        terminationId: number;
        subscriptionDate: Date;
        reSubscripedDate: Date;
        loanAct: string;
        hajjAct: string;
        persLoanAct: string;
        consumerLoanAct:string;
        otherAct1: string;
        otherAcc2: string;
        otherAcc3: string;
        otherAcc4: string;
        otherAcc5: string;
        empStreet1: string;
        empStreet2: string;
        cityCode: string;
        counCode: number;
        remarks: string;
        userId: string;
        activeDirectoryId: string;
        mainHrroleId: number;
        employeeLoginId: string;
        employeePassword: string;
        active: boolean;
        deleted: boolean;
        dateTime: Date;
        deviceId: string;
        uploadDate: Date;
        uploadby: string;
        syncDate: Date;
        syncby: string;
        synId: number;
        membership :string;
        membershipJoiningDate:Date;
        termination:string;
        terminationDate:Date;
        isKUEmployee: boolean;
        isOnSickLeave: boolean;
        isMemberOfFund: boolean;
        username:string;
    }